--move IBL instantiation out of RM constructor
Having the IBL env map loaded in the ResourceManager constructor was failing since there usually wasn't a GL context built yet, which was needed to instantiate the shader.

This is just a stop-gap measure so that IBL can be invoked successfully.  Much engineering needs to be done for IBL to be fully ready for prime time.